ImageResizing/requirements.txt
# Web framework
fastapi==0.68.0
uvicorn==0.15.0

# Database
pymongo==4.3.3
motor==3.1.1

# Data processing
pandas==1.5.3
numpy==1.21.6

# Image processing
Pillow==9.5.0
matplotlib==3.6.3

# Testing
pytest==7.2.2
pytest-asyncio==0.20.3
httpx==0.23.3
requests==2.28.2

# Utilities
python-multipart==0.0.5 